Refactor checkAuth as a utility library instead of action.
It only reads from the UserStore.
