+++ app/src/scripts/user/user.actions.js
@@ -1,7 +1,6 @@
 import Reflux from 'reflux'
 
 var Actions = Reflux.createActions([
-  'checkAuth',
   'getPreferences',
   'isRoot',
   'updatePreferences',
